feat:modify model, resources, refactoring etc...
