+++ .gitignore
@@ -1,0 +1,5 @@
+/packages/pue_ai/resources/weights/Emo_weights
+
+/packages/pue_ai/resources/weights/GD_weights
+
+/packages/pue_ai/resources/weights/NER_weights
