+++ app.py
@@ -2,15 +2,18 @@ from flask import Flask, jsonify
 from flask_restful import Api
 from flask_jwt_extended import JWTManager
 from flask_socketio import SocketIO
+from models import *
+from flask_bcrypt import Bcrypt
 from resources import create_api, create_socketio
-from resources.config.configure import config
+from flask_cors import CORS
 
 from db import db
 
+
+host = "127.0.0.1"
+port = 5001
 #SECRET_KEY = config['DEFAULT']['SECRET_KEY']
 #db_name = config['DEFAULT']['DB_NAME']+'.db'
-
-
 
 SECRET_KEY = "chan"
 db_name="chatbot"
@@ -27,12 +30,14 @@ app.config['PROPAGATE_EXCEPTIONS'] = True
 app.secret_key = "chan"
 api = Api(app) #API FLASK SERVER
 
-sock = SocketIO(app)
+sock = SocketIO(app,cors_allowed_origins="*")
 
 #this will be used for login(authenticate users)
 jwt = JWTManager(app) #this will make endpoint named '/auth' (username,password)
 #JWT will be made based on what authenticate returns(user) and JWT will be sent to identity to identify which user has Vaild JWT
+bcrypt = Bcrypt(app)
 
+# create_socketio(sock)
 #API works with resouce
 #200 ok
 #201 created
@@ -46,40 +51,39 @@ jwt = JWTManager(app) #this will make endpoint named '/auth' (username,password)
 #         return {'is_admin': True}
 #     return {'is_admin': False}
 #
-# @jwt.invalid_token_loader
-# def invalid_token_callback(error):  # we have to keep the argument here, since it's passed in by the caller internally
-#     return jsonify({
-#         'message': 'Signature verification failed.',
-#         'error': 'invalid_token'
-#     }), 401
-#
-# @jwt.unauthorized_loader
-# def missing_token_callback(error):
-#     return jsonify({
-#         "description": "Request does not contain an access token.",
-#         'error': 'authorization_required'
-#     }), 401
-#
-# @jwt.revoked_token_loader
-# def revoked_token_callback():
-#     return jsonify({
-#         "description": "The token has been revoked.",
-#         'error': 'token_revoked'
-#     }), 401
-create_api(api)
-create_socketio(sock)
+@jwt.invalid_token_loader
+def invalid_token_callback(error):  # we have to keep the argument here, since it's passed in by the caller internally
+    return jsonify({
+        'message': 'Signature verification failed.',
+        'error': 'invalid_token'
+    }), 401
+
+@jwt.unauthorized_loader
+def missing_token_callback(error):
+    return jsonify({
+        "description": "Request does not contain an access token.",
+        'error': 'authorization_required'
+    }), 401
+
+@jwt.revoked_token_loader
+def revoked_token_callback():
+    return jsonify({
+        "description": "The token has been revoked.",
+        'error': 'token_revoked'
+    }), 401
 
 @app.before_first_request
 def create_tables():
     db.create_all()
 
-
-
-
+create_api(api)
+create_socketio(sock)
 
 if __name__ == "__main__":
 
     db.init_app(app)
-    #app.run(port=3001,debug=True) #debug tells us what is problem
     print("Now we Run...")
-    sock.run(app,port=5000,debug=False)
+    #app.run(port=5000,debug=False) #debug tells us what is problem
+
+
+    sock.run(app,host=host,port=port,debug=False)
