--- packages/pue-chatbot/__init__.py
@@ -1,9 +1,0 @@
-from .setup import setup_environ, download_weights
-from .transformer_models.aimodel import AIModel
-__all__=['tensormodels','tokenizer', 'AIModel']
-
-setup_environ()
-download_weights()
-
-
-
