--- packages/pue-chatbot/resources/config.json
@@ -1,11 +1,0 @@
-{
-	"EMO-weights-version":"1.0.0",
-	"NER-weights-version":"1.0.0",
-	"Transformer-weights-version":"1.0.0",
-	"EMO-index-url":"https://drive.google.com/u/0/uc?id=1nSVbVcS75vc4Kf3XKOaXhLHMWlruvq95&export=download",
-	"EMO-data-url":"https://drive.google.com/u/0/uc?id=1wurKEIo8LaGv_AfaDrtXVrgnsXJ-HmMx&export=download",
-	"NER-index-url":"https://drive.google.com/u/0/uc?id=1fCK7UJMgwVd_eRkRSXZut2IHtGKVtTMV&export=download",
-	"NER-data-url":"https://drive.google.com/u/0/uc?id=1Q5HcqJLsWE5ZUzsnsSujQF0w8QHzsLTU&export=download",
-	"Transformer-index-url":"https://drive.google.com/u/0/uc?id=1lMc6Wx0baeC4VxI-3UuFXUCbw4OGVXXB&export=download",
-	"Transformer-data-url":"https://drive.google.com/u/0/uc?id=137jPMTIxu773g_AF4SPx2PjrzRGsOL_6&export=download"
-}
