--- packages/pue-chatbot/tensormodels/__init__.py
@@ -1,2 +1,0 @@
-from .emotion_recognizer import EmotionRecognizer
-__all__ = ['Encoding','WordClassifier', 'EmotionRecognizer']
