+++ packages/pue_ai/__init__.py
@@ -1,0 +1,4 @@
+from .setup import setup_environ, download_weights
+setup_environ()
+download_weights()
+
