+++ resources/chat.py
@@ -2,18 +2,8 @@ from flask_restful import Resource, reqparse
 from models.chat import ChatModel
 from flask_jwt_extended import jwt_required, get_jwt_identity
 
-class FullDateChat(Resource):
+class OneChat(Resource):
     parser = reqparse.RequestParser()
-    parser.add_argument('date_YMD',
-                        type=str,
-                        required=True,
-                        help="Field named 'date_YMD' cannot be blank."
-                        )
-    parser.add_argument('date_YMDHMS',
-                        type=str,
-                        required=True,
-                        help="Field named 'date_YMDHMS' cannot be blank."
-                        )
     parser.add_argument('direction',
                         type=str,
                         required=True,
@@ -24,23 +14,19 @@ class FullDateChat(Resource):
                         required=True,
                         help="Field named 'utterance' cannot be blank."
                         )
-    @jwt_required
-    def get(self, date):
-        user_id = get_jwt_identity()
-        chat = ChatModel.find_by_fulldate_with_user_id(user_id,date)
-        if chat:
-            return chat.json(), 200
-        return {'message': 'There is no chattings...'}, 404
+    parser.add_argument('emotion',
+                        type=str,
+                        required=True,
+                        help="Field named 'emotion' cannot be blank."
+                        )
 
-    @jwt_required
+    @jwt_required()
     def post(self,date):
         user_id = get_jwt_identity()
-        data = ChatModel.parser.parse_args()
 
-        if ChatModel.find_by_fulldate_with_user_id(user_id,date):
-            return {'message': "A chat with date '{}' already exists.".format(date)}, 400
+        data = OneChat.parser.parse_args()
 
-        chat = ChatModel(user_id, **data)
+        chat = ChatModel(user_id,date[:8],date[8:],data['direction'],data['utterance'])
 
         try:
             chat.save_to_db()
@@ -49,7 +35,7 @@ class FullDateChat(Resource):
 
         return chat.json(), 201
 
-    @jwt_required
+    @jwt_required()
     def delete(self, date):
         user_id = get_jwt_identity()
         chat = ChatModel.find_by_fulldate_with_user_id(user_id,date)
@@ -59,15 +45,22 @@ class FullDateChat(Resource):
         return {'message': 'Chat deleted'}, 200
 
 class AllChatList(Resource):
-    @jwt_required
+    @jwt_required()
     def get(self):
         user_id = get_jwt_identity()
         chats = [chat.json() for chat in ChatModel.find_all_by_user_id(user_id)]
         return {'chats': chats}
 
-class YMDChatList(Resource):
-    @jwt_required
-    def get(self,date):
+class RangeChatList(Resource):
+    @jwt_required()
+    def get(self,sday,eday):
         user_id = get_jwt_identity()
-        chats = [chat.json() for chat in ChatModel.find_all_by_dateYMD_with_user_id(user_id,date)]
+
+        return {'message': "Not implemented!"},505
+
+class YMDChatList(Resource):
+    @jwt_required()
+    def get(self,day):
+        user_id = get_jwt_identity()
+        chats = [chat.json() for chat in ChatModel.find_all_by_dateYMD_with_user_id(user_id,day)]
         return {'chats': chats}
